+++ core/Options.lua
@@ -256,38 +256,31 @@ function Options:Initialize()
 	layout:AddInitializer(CreateSettingsListSectionHeaderInitializer(L["options.about"]))
 
 	do
-		local data = {
+		layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_SettingsPanelTextNormal", {
 			leftText = L["options.about.game-version"],
 			rightText = PER.GAME_VERSION .. " (" .. PER.GAME_FLAVOR .. ")",
-		}
-
-		local text = layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_OptionsMenuText", data))
-
-		function text:GetExtent()
-			return 14
-		end
+		}))
 	end
 
 	do
-		local data = {
+		layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_SettingsPanelTextNormal", {
 			leftText = L["options.about.addon-version"],
 			rightText = PER.ADDON_VERSION .. " (" .. PER.ADDON_BUILD_DATE .. ")"
-		}
-
-		local text = layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_OptionsMenuText", data))
-
-		function text:GetExtent()
-			return 14
-		end
+		}))
 	end
 
 	do
-		local data = {
+		layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_SettingsPanelTextNormal", {
+			leftText = L["options.about.lib-version"],
+			rightText = AWL.ADDON_VERSION .. " (" .. AWL.ADDON_BUILD_DATE .. ")"
+		}))
+	end
+
+	do
+		layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_SettingsPanelTextLarge", {
 			leftText = L["options.about.author"],
 			rightText = PER.ADDON_AUTHOR
-		}
-
-		local text = layout:AddInitializer(Settings.CreateElementInitializer("ArcaneWizardLibrary_OptionsMenuText", data))
+		}))
 	end
 
 	do
@@ -296,7 +289,7 @@ function Options:Initialize()
 		local buttonText = L["options.about.button-github.button"]
 
         local function OnButtonClick()
-            AWL.Dialog:ShowLinkDialog(PER.LINK_GITHUB)
+            AWL.Dialogs:ShowLinkDialog(PER.LINK_GITHUB)
         end
 
         local buttonInitializer = CreateSettingsButtonInitializer(name, buttonText, OnButtonClick, tooltip, true)