+++ CHANGELOG.md
@@ -7,5 +7,7 @@
 **Important note: All settings have been reset to their defaults.**
 
 - Added: Required dependency 'Arcane Wizard: Library'
+- Changed: License
+- Updated: Addon description
 - Removed: ruRU localization
 - Refactoring of the addon structure and source code